UPDATE: build basic layout and implement components for code snippet viewer
To make custom imgui components easier to use, it's recommended to:
  1. Create a class subclassing `components.ImguiComponent`
  2. implement `render` function

Besides, note that `Vec2` in `components` module is a subclass of
`imgui.Vec2`. It's intended to make it easier to use with supports
of vector addition/subtraction and scalar multiplication. But
related tests should be done in the future.