UPDATE: implement helper class for resolving relations of nodes
`NodeCollection` is a helper class that can be used to resolve
relations of given nodes. This class will be useful to build data
which is required to be used in `CodeNodeViewer`.

Note that there are custom `__eq__()` methods in both `NodeLink`
and `NodeIndexLink` classes which make us able to compare links
even they are created dynamically instead of the original ones.