UPDATE: show context menu for adding node in `CodeNodeViewer`
Currently, either `imgui.begin_popup_context_window()` or
`imgui.begin_popup_context_item()` is failed to show the context
menu in the canvas ode `CodeNodeViewer`. And it's even failed to
work by creating a context menu manually according to the example
under the documentation of `imgui.core.begin_popup()`.

Therefore, we adopt a workaround to draw it. The main idea of this
workaround it drawing the context menu by the `DrawList` of window.
See also `CodeNodeViewer.create_context_menu()` for further details.