MAINT: rewrite code for showing context menu on the canvas of `CodeNodeViewer`
In previous commit 403cf73, we were not able to show context menu
on the canvas of `CodeNodeViewer`. But we figure out the cause now.

The reason why context menu is not able to be shown is that we did
not define a proper region as a "item" for popping up context menu.

To solve this issue, we can just define a "child" or "group" region
by `imgui.begin_child()` or `imgui.begin_group()` first, then call
either `imgui.begin_popup_context_item()` or
`imgui.begin_popup_context_window()` to show the context menu.

For a quick demonstration of the difference of defining a region,
consider these two snippets below:
1. without defining a region
    ```python
    imgui.begin()
    imgui.text('foo')
    imgui.text('bar')
    imgui.text('buzz')
    imgui.text('fizz')

    if imgui.begin_popup_context_item('context-menu', 2):
        imgui.selectable('menu-item-0')
        imgui.selectable('menu-item-1')
        imgui.end_popup()
    imgui.end()
    ```

2. with defining a region
    ```python
    imgui.begin()
    imgui.begin_group()
    imgui.text('foo')
    imgui.text('bar')
    imgui.text('buzz')
    imgui.text('fizz')
    imgui.end_group()

    if imgui.begin_popup_context_item('context-menu', 2):
        imgui.selectable('menu-item-0')
        imgui.selectable('menu-item-1')
        imgui.end_popup()
    imgui.end()
    ```

In the case 1, you can find that context menu will show up only when
the last text "fizz" is clicked.
In the case 2, context menu will show up no matter which text is
clicked.