MAINT: revise pattern for searching shared objects in Unix-like system
In Unix-like system, name of shared objects would be:
`[FILENAME].[PYTHON_VERSION_AND_PLATFORM].so`

So that the proper pattern for searching shared objects in forked
`pyimgui/imgui` repository should be `dir_imgui.rglob('core.*.so')`.

Besides, we should raise `ImportError` with different message to
indicate the actual cause of execution failure.