BUILD: replace `sys.argv` based approach for pre-installation with `setuptools.command.install` one
In order to make those built files of `pyimgui` able to be installed
to the site-package folder, we should use a custom command class
inheriting from `setuptools.command.install` instead.

Besides, we also have to generate a destination-source pairs as
the argument `data_files` for `setup()` command. See also
`prepare_data_files()` for details.

And the reason why `sys.argv` based approach implemented in previous
revision does not work for this case is that operations written in
`if ... in sys.argv: ...` are not actually executed before the
installation of this package itself.