BUILD: fallback to `sys.argv` based approach for installing forked `pyimgui`
There is a critical problem in previous implementation: value of
`data_files` are determined before running operations defined in
pre-intallation stage. Hence that `prepare_data_files()` does not
return desired result. (the reason why it worked previously is that
forked `pyimgui` repository was pulled already)

So that we fallback to `sys.argv` based approach. But we execute
those operations for pre-installation right before creating the
dictionary `metadata`.