BUG: "add_reference" event will be cancelled immediately when mouse cursor moves
This bug is introduced in commit 9fce6f1.

Since the facility of "add_reference" for nodes was implemented in
commit 10a794a, it relied on `imgui.set_window_focus_labeled()`
to bring user back to `CodeNodeViewer` and make them able to select
a node. And the event will be cancelled once user clicks elsewhere
(that is, event will be cancelled when `CodeNodeViewer` lost focus).

However, in commit 9fce6f1, window name of a `CodeNodeViewer` going
to be rendered is changed to the value `self.window_name` instead
of the constant string "CodeNodeViewer". But we forgot to change
the argument for `imgui.set_window_focus_labeled()`. And that's
why `CodeNodeViewer` won't be focused and resulted the issue
mentioned as the title of this commit.