UPDATE: highlight those nodes which can be added as leaf while "add reference" event is triggered
Criterions of recognizing a node which can be added as leaf are:
- "add reference" event is triggered
- a node doesn't have a root yet

When all the conditions are satisfied, a node will be highlighted
with a brighter border.

Besides, the circular reference check in `Node.add_leaf()` is
updated to check not only given `root` node, but also all the
`root` of root node. Otherwise, `NodeCollection.resolve_tree()`
will failed to find a valid tree since it have to find out nodes
without root before starting resolving leaf nodes.