MAINT: make node width fixed and show full name in a tooltip only when name is too long
In order to keep layout of nodes in canvas uniform, width of
`NodeComponent` is fixed and name of a node will be truncated if
it exceeds 8 characters (defined by `NODE_MAX_NAME_LENGTH`). But
user can still get the full name of those long-name nodes from the
ListView in `CodeNodeViewer` or just hover on a node.