BUG: states of modifier keys won't be reset after they are released
This is an issue in `pyimgui v1.2.0`, and it leads to the following
problem in this application:
- Collapsing header in `CodeSnippetWindow` cannot be opened once it
  is opened by "KEY_ALT + right click" on a `CodeNodeComponent`.
  Only when user presses a non-modifier key, collapsing header can
  be opened again.

- Once there is any single `CodeNodeComponent` is opened by "KEY_ALT
  + right click", all other `CodeNodeComponent`s can be opend by
  single right click instead of double-click.

We've been aware of this issue while developing callbacks for
`imgui.input_text_multiline()` widget in our forked version of
`pyimgui`. (See also [1]) But we had not figured out the real cause
of it.

The cause of this issue is revealed until we found the 2nd problem
mentioned above, becasue `imgui` is not responsible for handling
key stroke event. So, we started investigating possible modules
related to key event handling, and found that there are some problems
in `PygletMixin.on_key_press()` and `PygletMixin.on_key_press()`.

More details about solving this issue can be found in docstrings in
file `_patches.py`.

[1]: https://github.com/NaleRaphael/pyimgui/blob/af565243/imgui/core.pyx#L4923-L4927