MAINT: avoid marking selected node itself as a node which can be added as leaf reference
In commit 0c2fa6f, a feature for highlight those nodes which can
be added as leaf reference was introduced. However, the node
selected as a root itself was not excluded to be highlighted.

To solve this issue, we should also checking whether a node is
currently a selected node (of which the index will be recorded
in `CodeNodeViewer.id_selected`).