MAINT: keep opened `CodeSnippetWindow` open after rearranging nodes
In previous revisions, opened `CodeSnippetWindow` will be closed
if nodes are rearranged. The cause of this issue is that
`node_components` in `CodeNodeViewer` will be recreated according
to resolved tree generated from `node_collection`. Since those
`CodeNodeComponent`s are newly created, state of `CodeSnippetWindow`
cannot be kept.

To solve this issue, we should instantiate `CodeNodeComponent`s
only when a `CodeNodeViewer` is created. If `node_components` is
not empty and function for node rearrangement is invoked, we should
update position of `node_components` rather than re-create them.