MAINT: adjust inital settings for `CodeSnippetWindow`
In `CodeSnippetWindow`, width of snippet region was only up limited.
So content of snippet would be truncated if the longest line of it is
still too short. To solve this, variable for minimal width is also
added to make sure code won't be truncated.

Besides, those code for calculating these layout related variable is
moved into a function `calculate_window_size()`. So that we can also
re-calculate these settings after editing content of snippet, and make
the content region still fit.

In addition, header in the snippet content are now removed.