MAINT: add a new key mapping to detect KEY_TAB when KEY_SHIFT is pressed
While we were developing extension methods for paragraph indentation,
we found that KEY_TAB cannot be detected when KEY_SHIFT is pressed.
And it was considered as an issue of `imgui` package. However, since
we found that `imgui` is not responsible for hanlding key stroke
event, we found that it's actually an issue comes from an integrated
module `pyglet` in `pyimgui` and made a pacth for it (see also 6ae07b5)

But this issue still exist after adpoting that patch. After investigating
the incoming key symbol and modifiers, we found that incoming symbol
will be an strange number (98784247808) which hasn't been listed in
offical package (`pyglet.window.key`). In order to make KEY_TAB able
to be detected when KEY_SHIFT is pressed, we add a new key mapping
for this number.

Currently we still cannot figure out why key symbol becomes to this
number, so that we may need to keep this issue in mind in case there
is another issue related to key stroke handling.