MAINT: rewrite event system
Name-resolution based event system is now rewritten into the one
using "event_name -> subscriber_method". The problem of previous
implementation is that it leads to a constraint of naming a
method as a handler for specific event. Besides, it makes
components have to handle the event registration even if there
is no related event handler in itself (see also [1] and [2]).

To solve this design problem, the new event system replys on a
singleton event registry `NodeEventRegistry`. It's responsible
for mapping subscribers to corresponding event. Those components
utilizing it are just need to keep a reference of this registry
and register handler for event during the initialization of
themselves.

But note that components are also responsible for unregistrying
handlers before closing/terminating themselves. See also the
implementation of `CodeNodeViewer.close()` for further details.

[1]: https://github.com/NaleRaphael/codememo/blob/9f934b5/codememo/components.py#L625
[2]: https://github.com/NaleRaphael/codememo/blob/9f934b5/codememo/components.py#L627-L630