BUG: fix issue of incorrectly highlighting nodes as ones can be added as leaf reference
Those nodes which already have roots were not excluded from the
candiates to be added as leaf reference. This condition is enough
for us to exclude most of nodes that should not be a candidate.

Besides, we should also check whether a node is the root in upper
hierarchy of the one going to create leaf reference.