MAINT: add window ID to prevent window-based widget being duplicated
Since `imgui` is using ID stack to manage windows, content of
windows with the same ID will be rendering in the same window.
This is not a bug, but since there are some components shouldn't
behave in this way, we should suffix their names with IDs.

Besides, there is also an issue of using just event name as a key
for event system. This will be problematic if there are multiple
instances of widgets that require to handling the same event.
e.g. A node will be created in two `CodeNodeViewer`s because they
register their own handler will the same event name "create_node".

To solve this problem, name of event to be registered and fired
are now suffixed with the ID of `CodeNodeViewer`.

Other revisions:
- `CodeNodeViewer` and `CodeNodeCreatorWindow` now use a time based
  ID as their window identifiers in order to prevent duplication.

- `CodeSnippetPropertyWindow` now use UUID of a node to prevent
  duplication.

- Timing of instantiating `CodeSnippetWindow` is changed. It was
  going to be created when a `CodeNodeComponent` is instantiated,
  but now it will be created only when user double-clicked (or
  click + KEY_ALT) a `CodeNodeComponent`.

- In `CodeSnippetWindow`, a new argument `container_id` is added
  to `__init__()` because it requires to fire the event
  "add_reference" with the ID of container it belongs to. This
  approach is currently a solution for the above-metioned problem.

- Naming style of names used in `imgui.begin_child()` are now all
  changed to hyphenated words.