MAINT: add event handler for holding down keys
Handler `on_text_motion()` is added to patched `PygletMixin` in order
to pass events related to holding down keys. Without this handler,
we cannot move a text curosr (caret) repeatedly by holding arrow keys.

Currently we just need to add this handler to the handler stack of
pyglet window without adding implementation in it. But the latency
of the beginning of text cursor movement while holding arrow keys
is a bit high, we may need to improve this.
(result of current investigation: this issue is not highly relevant
to event loop handled by `pyglet` since speed of invocation of this
handler is faster than the movement speed of text cursor)