UPDATE: implement shortcut for main menu bar
Because shortcut API has not been implemented in `imgui`, passed
`shortcut` argument in `pyimgui.imgui.menu_item()` is currently
just a display option (see also issue 456 in `ocornut/imgui`).

But since key states have been handled in both `imgui` and
`pyimgui[pyglet]` now, we can implement a simple system for detecting
shortcut.

The idea of our approach for this topic is simple but it relies
of our current application architecture, here is a workflow of it
(see also `update()` in `Application.init_draw_process()`):
1. When there is a new frame created, we poll all key combinations
  of registered shortcuts, and cache state of triggered shortcuts.
2. During the process of rendering other imgui components, every
  component can access a registry for shortcut attached under the
  instance of the main application, and check whether their registered
  shortcuts have been triggered.
3. At the end of the process of rendering other imgui components,
  cached state of triggered shortcuts will be cleared from registry.

But note that the currently implemented `ShortcutRegistry` works
globally, so that developers should be careful of binding collisions.

And since `imgui` only stores key pressing state (in `ImGuiIO.KeysDown`
array), we cannot distinguish the difference between key-pressing
and key-holding events throuh it. So currently we have to check
whether a shortcut should be edge-triggered or not. Related mechanism
for this is implemented in `poll()` and `clear()` in `ShortcutRegistry`,
see also them for further details.