UPDATE: implement shortcut for `CodeNodeViewer`
Shortcuts for the following functionalities are added:
- Save (save project)
- Save As (save existing project to a new file)

But since `CodeNodeViewer` is a widget that there can be multiple
instances existing simutaneously, we have to check whether shortcuts
have been registered every time a `CodeNodeViewer` instance is
created. Though this approach works currently, we may need to figure
out a better solution in the future.