Implement shortcut for menu items (#3)
* MAINT: handle key states of ASCII characters

Key state of ASCII characters are not handled in `pyimgui[pyglet]`
v1.2.0. Without handling these key states, we won't be able to
implement shortcuts for menu items.

Except setting key states of ASCII characters, we also have to
deal with those special keys related to text edit (e.g. CTRL+A,
CTRL+C, CTRL+V, ...). Since the argument `symbol` passed to callbacks
`on_key_press()` and `on_key_release()` are the same no matter
key CTRL is pressed or not, we have to modify the original
implementation of these 2 methods in order to handle these cases.
See also the modified version of these 2 methods for more details.

* UPDATE: implement shortcut for main menu bar

Because shortcut API has not been implemented in `imgui`, passed
`shortcut` argument in `pyimgui.imgui.menu_item()` is currently
just a display option (see also issue 456 in `ocornut/imgui`).

But since key states have been handled in both `imgui` and
`pyimgui[pyglet]` now, we can implement a simple system for detecting
shortcut.

The idea of our approach for this topic is simple but it relies
of our current application architecture, here is a workflow of it
(see also `update()` in `Application.init_draw_process()`):
1. When there is a new frame created, we poll all key combinations
  of registered shortcuts, and cache state of triggered shortcuts.
2. During the process of rendering other imgui components, every
  component can access a registry for shortcut attached under the
  instance of the main application, and check whether their registered
  shortcuts have been triggered.
3. At the end of the process of rendering other imgui components,
  cached state of triggered shortcuts will be cleared from registry.

But note that the currently implemented `ShortcutRegistry` works
globally, so that developers should be careful of binding collisions.

And since `imgui` only stores key pressing state (in `ImGuiIO.KeysDown`
array), we cannot distinguish the difference between key-pressing
and key-holding events throuh it. So currently we have to check
whether a shortcut should be edge-triggered or not. Related mechanism
for this is implemented in `poll()` and `clear()` in `ShortcutRegistry`,
see also them for further details.

* MAINT: handle shortcuts of which the key combinations are partially equal

In previous revision, multiple shortcuts will be triggered if their
key combinations are partially equal. e.g. Pressing `CTRL+SHIFT+S`
will make shortcuts with key combination of `CTRL+S` and `CTRL+SHIFT+S`
be triggered.

In order the solve this issue, triggered shortcut validated by
`is_triggered()` will be stored in to a list `candidates` first, then
we will select the one with the longest key combination as the
actually triggered shortcut.

* MAINT: rewrite `handle_shortcut()` in `MenuBar` widget

In `ShortcutRegistry`, `triggered_shortcuts` was changed from a
set to a string value because it allows only one shortcut to be
triggered at a time. So that releated code in `handle_shortcut()`
is rewritten.

* UPDATE: implement shortcut for `CodeNodeViewer`

Shortcuts for the following functionalities are added:
- Save (save project)
- Save As (save existing project to a new file)

But since `CodeNodeViewer` is a widget that there can be multiple
instances existing simutaneously, we have to check whether shortcuts
have been registered every time a `CodeNodeViewer` instance is
created. Though this approach works currently, we may need to figure
out a better solution in the future.