MAINT: show line number based on the first line from refered snippet
In previous revisions, column width of line number won't fit to the
total lines of code snippet. That's because we are using the line
number from refered code snippet to calculate proper column width
for line number in the table of `CodeSnippetWindow`. To fix it, we
just have to display those line number based on refered snippet
instead of the total lines of given snippet.

But we might need to add an option to let user able to select the
mode of line number display.