MAINT: clear reference of file_dialog after it's terminated
In commit cdb69b1, `file_dialog` in `MenuBar` would be checked before
creating a new `FileDialog` instance in order to prevent duplicate.
However, reference hold by it wouldn't be cleared after `FileDialog`
is closed by clicking "X" button at the top-right corner of window.
And it leads to `FileDialog` won't be created after that.

To fix this issue, we should check whether state `terminated` of
`FileDialog` in every new frame and clear the reference once it's
true.