MAINT: handle recursive function calls while parsing a DOT file for call graph, and add serial numbers to nodes generated from multi-root node
Since self-reference node is not allowed in our implementation, we
cannot create a link to represent recursive function calls. So that
we add a suffix "(recursive)" into the function name when a function
is detected to call itself recursively.

Besides, those duplicate nodes generated by multi-root node (a function
which is invoked by multiple functions) are now suffixed by a serial
number to make them easier to be identified.