BUILD: change option of using forked pyimgui to a default one, and update README for intalling `pygraphviz` on non-linux machine
Forked version of `pyimgui` is now installed by default since those
custom functionalities are important for our application. But users
can still choose to use original version of it by specifying a install
option into pip command: `--global-option="--use-original-pyimgui"`

Besides, since `pygraphviz` requires source of `graphviz` to build
and there would be some complicated settings have to be done on
non-linux machines, it's removed from the list of `extras_require`
and will be added back only when a machine is detected as linux.

As for non-linux users, `graphviz` and `pygraphviz` has to be
downloaded and installed manually. But it will be much easier to
install a forked version through `conda`. These instructions are
documented in README.