BUG: Rewrite code for finding root and orphan nodes
Major updates/revisions:
The original implementation of the code for finding root and orphan
nodes will fail when nodes are chaining and forming a single circular
reference loop, e.g.
```
A -> B -> C
 \---<---/
```
In this cases, no orphans and roots will be found out. As a result,
no tree will be built. However, it's clearly wrong because there is
actually an exisiting relation (i.e. (A -> B, B -> C, C -> A)).

To solve this issue, we pick a random node as root when no global
root is found, then use that root to traverse possible trees. And we
repeat this process until all nodes are visited. See also that local
function `find_roots_and_orphans()` in `NodeCollection.resolve_tree()`
for further details.

Minor updates/revisions:
- In method `objects.Node.remove_leaf_by_index()`, `node.reset_root()`
  requires a parameter.
- In `test_objects.py`, comparisons of node links are all replaced by
  set comparison instead of list comparison to prevent them being flaky
  cases.