Change node structure from a single-root one to a multi-root one  (#6)
* MAINT: Change node structure from a single-root one to a multi-root one

Though single-root nodes are a common used structure for building trees,
trees formed by single-root nodes are not always able to well represent
relations of functions, e.g. recursion function calls and a function
that can be invoked by multuple functions.

And as it's stated in PR #4, imported call graph have to be modified
slightly in order to convert those multi-root/self-referenced nodes into
single-root nodes. Though it won't modify the relation of functions,
those duplicate nodes may misinform the actual relation between functions.

In order to improve it, there are serveral changes done in this commit.

Major updates/revisions:
- In `objects.Node` class:
  - Field `root` is changed to `roots`, and it takes a list of nodes
    as values.
  - Filed `_ref_info` is removed and property `ref_info` is now changed
    to `ref_infos` which is a mapping of "uuid of a root node" to
    "an instance of `ReferenceInfo`".
  - Since there might be multiple root nodes now, method `reset_root()`
    is modified to take an additional parameter `node`.
  - In method `add_leaf()`, exceptions of "Multiple root", "Self
    reference" and "Circular reference" are removed because these
    situations are allowed in call graphs.

- In `objects.NodeCollection` class:
  - Since there might be multiple root nodes now, an additional
    parameter `root` is added into method `remove_root_reference()`.
  - In method `resolve_links()`, `resolve_links_from_trees()` and
    `resolve_index_links()`, parameter `root_slot` of `NodeLink()`
    is not always 0 anymore.
  - In method `resolve_tree()`, an additional parameter `visited` is
    added into local function `build_tree()` in order to record
    visited nodes and avoid duplicate.

- In `components` module, all implementations related to `Node.root`
  are revised since the new field `Node.roots` is always a list.

Minor updates/revisions:
- In method `graph_parser.DotParser.parse()`, process for converting
  multi-root nodes into single-root nodes are now removed.
- In `CodeNodeComponent.render()`,  method
  `check_node_can_be_added_as_reference()` is removed since there is
  no more restrictions of adding a leaf reference.
- Broken test cases related to these changes are also revised.

* TST: remove node names with serial number

Those node names with serial number are created to indicate a
function is invoked by multiple functions. But since multi-root
node structure are now supported, there is no need to rely on
this approach anymore.

* BUG: Rewrite code for finding root and orphan nodes

Major updates/revisions:
The original implementation of the code for finding root and orphan
nodes will fail when nodes are chaining and forming a single circular
reference loop, e.g.
```
A -> B -> C
 \---<---/
```
In this cases, no orphans and roots will be found out. As a result,
no tree will be built. However, it's clearly wrong because there is
actually an exisiting relation (i.e. (A -> B, B -> C, C -> A)).

To solve this issue, we pick a random node as root when no global
root is found, then use that root to traverse possible trees. And we
repeat this process until all nodes are visited. See also that local
function `find_roots_and_orphans()` in `NodeCollection.resolve_tree()`
for further details.

Minor updates/revisions:
- In method `objects.Node.remove_leaf_by_index()`, `node.reset_root()`
  requires a parameter.
- In `test_objects.py`, comparisons of node links are all replaced by
  set comparison instead of list comparison to prevent them being flaky
  cases.

* MAINT: Rewrite code for interaction of removing root reference

Since we supported building a tree/call graph with multi-root nodes,
an additional confirmation should be done when users want to remove
root reference of a multi-root node.

Now, roots of a selected multi-root node will be highlighted by
red border when "Remove root reference" evnet is triggered. And this
process will continue only when user click a valid root node.

* MAINT: remove last empty layer when there is a circular reference loop in tree

* TST: add more tests for `objects` module and add a check statement into `Node.remove_leaf()`

* MAINT: rewrite `NodeCollection.resolve_index_links_from_trees()`

Flatten tree can be built outside of the loop, and we can omit
the counter `prev_tree_size` by adopting this approach.
Besides, parameter `root_slot` for creating `NodeIndexLink is
revised since multi-root node structure are supported now, and it
won't be always 0.

* MAINT: rename `resolve_tree()` to `resolve_trees()`

Outputs returned by `NodeCollection.resolve_tree()` are trees and
orphans, so that it should be renamed as `resolve_trees()` to avoid
confusion.

* MAINT: remove unused variable

`multi_root_counter` was used for counting numbers of multi-root
nodes and generating serial number for duplicate nodes while building
trees from a call graph. But since multi-root nodes are now supported,
there is no need to handle this case anymore.