BUILD: fix path returned by custom `getsitepackages()`
In PR #5, a custom method `getsitepackages()` was implemented in
order to solve the issue of installing under `virtualenv` on
Travis-CI. But that path was not configured correctly, it was set
to the folder of `pip` rather than its parent directory (i.e.
site-package / dist-package).

See also the build history in this link:
https://travis-ci.com/github/NaleRaphael/codememo/jobs/411530090#L1281-L1299