BUILD: replace command `python -m pytest` with `pytest` to run tests
"codememo/vendor/settings.py" can only be generated after running
installation command, so that it won't exist in source code. But
since tests were run by command `python -m pytest` on Travis CI,
the local `codememo` package has a high priority to be imported.

It makes test runner not able to find "codememo/vendor/settings.py"
and lead to the warning message "settings.py are not available"
being issued.

To solve this issue, we can just replace the command `python -m pytest`
with `pytest`. With the latter one, local path won't be included,
so that we can import the correct package which is installed in
site-package.