BUILD: switch back to use `python -m pytest` as test command
Though running test by command `pytest` can avoid issue resulted
by including path of local package, it makes `coverage.py` not able
to collect coverage data from test cases.

The reason why `coverage.py` failed is that "tests" folder has to
be a module when we are running `pytest` as test command. But since
there is no "__init__.py" in "tests" folder, it won't be detected as
a valid Python module. So that `coverage.py` failed to collect data.
See also link [1].

This issue can be solved by adding "__init__.py" files into "tests"
folder, as it's mentioned in pytest documentation [2]. But it also
introduce a new problem (also stated in [2]):

> in order to load the test modules from the tests directory, pytest
> prepends the root of the repository to sys.path, which adds the
> side-effect that now mypkg is also importable.

Besides, we prefer not to make "tests" importable. So that we decide
to revert the changes made in 0be6e6. As for the warning issued by
missing "codememo/vendor/settings.py" file, it won't be a problem
currently. Because the building process of Cython extension will be
run while installing this package, and we can checkout whether there
is a problem if installation failed. Besides, current test cases are
mainly made for other modules excluding `codememo.components`.

We can continue solving this issue after test framework for `imgui`
and pytest 6.0 are released. (note: In pytest 6.0,
"--import-mode=importlib" can be used to solve this problem)

[1]: https://stackoverflow.com/a/60579142
[2]: https://docs.pytest.org/en/stable/goodpractices.html#tests-outside-application-code