MAINT: close `CodeNodeCreatorWindow` after `CodeNodeViewer` is closed
Both `CodeNodeCreatorWindow` and `CodeNodeViewer` are rendered by
the main render loop (i.e. the local function `update()` in
`Application.init_draw_process()`), so that they can live until
the whole application is terminated. However, what we expected is
`CodeNodeCreatorWindow` should be closed when its container is
closed.

To solve this issue, a field `terminated` is added in `CodeNodeViewer`,
and it will be set to True after a `CodeNodeViewer` is closed. So
that `CodeNodeCreatorWindow` can check it before starting rendering
and close itself after that flag is set to True.